% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bray_diss.R
\name{bray_diss_z}
\alias{bray_diss_z}
\title{Dissimilarity z-score}
\usage{
bray_diss_z(x, B = 500)
}
\arguments{
\item{x}{a site by species matrix as used by \code{vegan::vegdist}}

\item{B}{number of permutation replicates}
}
\value{
a single numeric value representing the z-score
}
\description{
Calculate z-score of Bray-Curtis dissimilarity compared to null model
generated with quasi-swap algorithm
}
