% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bray_diss.R
\name{bray_part}
\alias{bray_part}
\alias{bray_part_z}
\title{Partition Bray-Curtis dissimilarity}
\usage{
bray_part(x, grp)

bray_part_z(x, grp, B = 1000)
}
\arguments{
\item{x}{a site by species matrix as used by \code{vegan::vegdist}}

\item{grp}{logical, \code{TRUE} for one group and \code{FALSE} for the other}
}
\value{
a single number, the log of the ratio of contributions of each group to
the overall dissimilarity; a negative number means a greater contribution 
of the group represented by \code{grp = FALSE}, a positive number means a 
greater contribution of the group represented by \code{grp = TRUE}.

\code{bray_part_z} returns a permutation-based z-value (using \code{cswap})
of the same log ratio. negative and positive sign of z-value is same as for 
\code{bray_part} but is now a "greater contribution of the group ..., 
relative to chance."
}
\description{
Partition Bray-Curtis dissimilarity across two different groups of species
}
\details{
only works for a logical \code{grp} though the same partitioning logic 
applies to greater than two groups
}
