% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bray_diss.R
\name{cswap}
\alias{cswap}
\title{Sequential column permutation algorithm}
\usage{
cswap(x, g)
}
\arguments{
\item{x}{a site by species matrix as used by \code{vegan::vegdist}}

\item{g}{logical, \code{TRUE} for one group and \code{FALSE} for the other}
}
\value{
permuted group vector of \code{TRUE}s and \code{FALSE}s ready to be used in
\code{bray_part}
}
\description{
Create a randomization of column groupings using a sequential algorithm
}
\details{
seeks to probabilistically constrain column sums within groups meaning
column sums will not be strictly preserved, but the distributed of column
sums within groups should be similar between observed and permuted data. 
the algorithm does this by sequentially swapping columns with probability
inversely proportional to the relative magnitude of difference between
column totals
}
