% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_mete.R
\name{samp_mete}
\alias{samp_mete}
\title{Sample from a \code{meteDist} object}
\usage{
samp_mete(x, n)
}
\arguments{
\item{x}{the \code{meteDist} object}

\item{n}{number of random samples}
}
\value{
a list, usually of length \code{n} with each element containing a 
sample. if length is less than \code{n} a warning is printed.
}
\description{
A function to draw random samples from a \code{meteDist} object while keeping
the state variables of the samples and original data in reasonable
agreement
}
