% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_means.R
\name{site_means_lm}
\alias{site_means_lm}
\title{Site means meta regression}
\usage{
site_means_lm(formula, data)
}
\arguments{
\item{formula}{a formula where random effects notation is used to indicate
grouping variable (i.e. site)}

\item{data}{the data frame where variables in the formula live}
}
\value{
a \code{metafor::rma} object
}
\description{
run a meta regression on raw data. group means will be computed internally
}
\details{
only works when provided both a formula and data. formula must include a
grouping term in random effects style syntax, \code{(1 | group)}. only a 
fixed effects rma is calculated, i.e. \code{metafor::rma(..., method = "FE")}
}
