% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_means.R
\name{site_means_predict}
\alias{site_means_predict}
\title{Predict site means from meta regression}
\usage{
site_means_predict(mod, newdata)
}
\arguments{
\item{mod}{a \code{metafor::rma} object}

\item{newdata}{optional data frame with values for explanatory variable(s)
column names must match those used when making \code{mod}}
}
\value{
a \code{metafor::predict.rma} object, similar to a data frame or list with 
elements for \code{pred}, \code{se}, \code{ci.lb}, and \code{ci.ub}
}
\description{
predict response and confidence intervals from a \code{metafor::rma} object
}
\details{
assumes a model with an intercept term. assumes \code{newdata} has already
been grouped and summarized by site (i.e. rows represent group means)
}
