% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samp_mete.R
\name{summarize_samp_mete}
\alias{summarize_samp_mete}
\title{Summarize output from \code{samp_mete}}
\usage{
summarize_samp_mete(x)
}
\arguments{
\item{x}{output of \code{samp_mete}}
}
\value{
a data frame with columns \code{rank}, \code{m} for mean, and 
\code{lo} and \code{hi} for 95% CI bounds
}
\description{
create a data frame with mean and 95% CI calculated for each rank
across samples produced by \code{samp_mete}
}
\details{
if the upper bound of the CI collapses to 1 then a small wiggle is added
so the upper and lower bounds remain distinguishable
}
