% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this_ms_look.R
\name{this_ms_look}
\alias{this_ms_look}
\title{Custom plotting parameters}
\usage{
this_ms_look(log_x = FALSE, log_y = FALSE)
}
\arguments{
\item{log_x}{boolean, if the x-axis is to be logged, default is FALSE}

\item{log_y}{boolean, if the y-axis is to be logged, default is FALSE}
}
\description{
A set of plotting choices to keep consisten plots across the project
}
